Use `ROM::Initializer` in `CSV::Gateway`
Updates the gateway to use the ROM initializer class, which is a thin
wrapper around `dry-initializer` - eliminating some boilerplate and
allow for some type enforcement for params.

Of note, the `options` param was renamed to `csv_options` since
`dry-initializer` *did not like* a param named `options`. Also, the
`datasets` attr reader was moved up to make it a little more obvious.